requirements.txt
aiosqlite
fastapi
greenlet
pika
python-dotenv
redis
SQLAlchemy
uvicorn
